+++ sdk/resourcemanager/synapse/armsynapse/zz_generated_pagers.go
@@ -1798,6 +1798,60 @@ func (p *WorkspaceManagedSQLServerBlobAuditingPoliciesClientListByWorkspacePager
 	return p.current
 }
 
+// WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListPager provides operations for iterating over paged responses.
+type WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListPager struct {
+	client    *WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClient
+	current   WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListResponse
+	err       error
+	requester func(context.Context) (*policy.Request, error)
+	advancer  func(context.Context, WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListResponse) (*policy.Request, error)
+}
+
+// Err returns the last error encountered while paging.
+func (p *WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListPager) Err() error {
+	return p.err
+}
+
+// NextPage returns true if the pager advanced to the next page.
+// Returns false if there are no more pages or an error occurred.
+func (p *WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListPager) NextPage(ctx context.Context) bool {
+	var req *policy.Request
+	var err error
+	if !reflect.ValueOf(p.current).IsZero() {
+		if p.current.DedicatedSQLminimalTLSSettingsListResult.NextLink == nil || len(*p.current.DedicatedSQLminimalTLSSettingsListResult.NextLink) == 0 {
+			return false
+		}
+		req, err = p.advancer(ctx, p.current)
+	} else {
+		req, err = p.requester(ctx)
+	}
+	if err != nil {
+		p.err = err
+		return false
+	}
+	resp, err := p.client.pl.Do(req)
+	if err != nil {
+		p.err = err
+		return false
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		p.err = runtime.NewResponseError(resp)
+		return false
+	}
+	result, err := p.client.listHandleResponse(resp)
+	if err != nil {
+		p.err = err
+		return false
+	}
+	p.current = result
+	return true
+}
+
+// PageResponse returns the current WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListResponse page.
+func (p *WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListPager) PageResponse() WorkspaceManagedSQLServerDedicatedSQLMinimalTLSSettingsClientListResponse {
+	return p.current
+}
+
 // WorkspaceManagedSQLServerEncryptionProtectorClientListPager provides operations for iterating over paged responses.
 type WorkspaceManagedSQLServerEncryptionProtectorClientListPager struct {
 	client    *WorkspaceManagedSQLServerEncryptionProtectorClient